# backend/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Internationalized culture / e-commerce / custom-travel platform for Jingdezhen ceramic art.\nBackend in Go + Fiber; money as BIGINT minor units (CNY base, USD/EUR/GBP presentment).",
        "title": "Jingdezhen Ceramics Platform API",
        "termsOfService": "https://jingdezhen.test/terms",
        "contact": {
            "name": "Jingdezhen Ceramics Platform Team"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/admin/media/assets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Paginated list of all media_assets, optionally filtered by kind.\nAccess: content_editor.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "media"
                ],
                "summary": "List media assets (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccess_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter by kind (image|video)",
                        "name": "kind",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (1-based)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size (max 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{data: []models.MediaAsset, page, limit, total}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (needs content.write)",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Records an already-uploaded file as a media_assets row (after\nthe browser has uploaded via presign or local upload). Access: content_editor.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "media"
                ],
                "summary": "Register a media asset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccess_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Asset metadata (oss_key, mime, kind, dimensions)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.RegisterAssetData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.MediaAsset"
                        }
                    },
                    "400": {
                        "description": "Invalid body / validation / invalid media kind",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (needs content.write)",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/media/assets/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes a media_assets row (and the underlying object in OSS mode).\nAccess: content_editor.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "media"
                ],
                "summary": "Delete a media asset (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccess_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Asset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content (empty body)"
                    },
                    "400": {
                        "description": "Invalid id",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (needs content.write)",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Asset not found",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/media/presign": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a presigned PUT URL (OSS mode) or a local upload URL\n(local mode) for the browser to upload a media file to.\nAccess: content_editor (content.write).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "media"
                ],
                "summary": "Presign a media upload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccess_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "{kind: image|video, mime: string, size: int}",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{mode: oss|local, upload_url, oss_key, headers, public_url?}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid body / validation",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (needs content.write)",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/catalog/products/{id}/media": {
            "get": {
                "description": "Returns a product's ordered gallery for the storefront. No auth.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catalog",
                    "products",
                    "media"
                ],
                "summary": "List a product's media gallery (public)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ProductMediaItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid product id",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "jingdezhen-ceramics-backend_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "Optional additional details",
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "jingdezhen-ceramics-backend_internal_models.RegisterAssetData": {
            "type": "object",
            "required": [
                "kind",
                "mime",
                "oss_key"
            ],
            "properties": {
                "duration": {
                    "description": "videos",
                    "type": "integer",
                    "minimum": 0
                },
                "height": {
                    "type": "integer",
                    "minimum": 0
                },
                "kind": {
                    "type": "string",
                    "enum": [
                        "image",
                        "video"
                    ]
                },
                "mime": {
                    "type": "string"
                },
                "oss_key": {
                    "type": "string"
                },
                "width": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer \u003caccess_token\u003e\" — JWT access token from POST /auth/login (or OAuth). Required for all /admin/* and customer-authed endpoints; omitted for public /catalog/* reads.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}